+++ React/project_arduino/src/components/Form/RolForm.jsx
@@ -6,13 +6,9 @@ import { TitleComponent } from "../Title/TitleComponent";
 import UpdateIcon from "@mui/icons-material/Update";
 import CloseIcon from "@mui/icons-material/Close";
 import AddIcon from "@mui/icons-material/Add";
-import { apiUrl, port } from '../../utils/config';
+import { apiUrl, port } from "../../utils/config";
 import { NavAdmin } from "../Layouts/NavBar/NavBarAdmin/NavBar";
-import {
-  handleInputChange,
-} from "../../utils/validate";
-
-
+import { handleInputChange } from "../../utils/validate";
 
 export const RolForm = () => {
   const [forms, setForm] = useState([]);
@@ -23,11 +19,10 @@ export const RolForm = () => {
 
   const [currentPage, setCurrentPage] = useState(0); // Estado para el número de página actual
   const PerPage = 6; // Número de roles por página
-  const form = 'rols';
-  const title = useRef("Administrador de roles") //no actualiza la pagina en cambio el state si 
+  const form = "rols";
+  const title = useRef("Administrador de roles"); //no actualiza la pagina en cambio el state si
 
   const handleGet = async () => {
-    
     try {
       const response = await fetch(`${apiUrl}${port}/${form}`);
       const data = await response.json();
@@ -36,7 +31,7 @@ export const RolForm = () => {
       console.error("Error al obtener roles:", error);
     }
   };
-  
+
   useEffect(() => {
     // Actualizar la tabla cuando se actualiza el estado 'forms'
     handleGet();
@@ -46,36 +41,42 @@ export const RolForm = () => {
   useEffect(() => {
     // Actualizar la tabla cuando se actualiza el estado 'forms'
     handleGet();
-    setDeleted(false)
+    setDeleted(false);
   }, [deleted]);
-  
+
   useEffect(() => {
     // Actualizar la tabla cuando se actualiza el estado 'forms'
     handleGet();
-    setDeleted2(false)
+    setDeleted2(false);
   }, [deleted2]);
 
   const handleSubmit = async (e) => {
     e.preventDefault();
-    
+
     if (!news.description_rol) {
       //funciona en el caso del que el campo no sea required
       alert("Por favor, ingrese una descripción válida.");
       return;
     }
-  
+
     if (selected) {
       if (window.confirm("¿Estás seguro de que quieres actualizar este rol?")) {
         try {
           // Enviar solicitud para actualizar un rol existente
-          const response = await fetch(`${apiUrl}${port}/${form}/${selected.id_rol}`, {
-            method: "PUT",
-            headers: {
-              "Content-Type": "application/json",
-            },
-            body: JSON.stringify(news),
-          });
+          const response = await fetch(
+            `${apiUrl}${port}/${form}/${selected.id_rol}`,
+            {
+              method: "PUT",
+              headers: {
+                "Content-Type": "application/json",
+              },
+              body: JSON.stringify(news),
+            }
+          );
           const data = await response.json();
+          data.error == "Ocurrió un error en el servidor"
+            ? alert("Ya existe otro rol llamado igual.")
+            : alert(data.message);
           setForm((prev) =>
             prev.map((rol) => (rol.id_rol == data.id_rol ? data : rol))
           );
@@ -97,7 +98,10 @@ export const RolForm = () => {
         });
         const data = await response.json();
 
-        data.error == "Ocurrió un error en el servidor" ?  alert("ya existe") : setForm((prev) => [...prev, data]); setNew({ description_rol: "" });
+        data.error == "Ocurrió un error en el servidor"
+          ? alert("Ya existe otro rol llamado igual.")
+          : setForm((prev) => [...prev, data]);
+        setNew({ description_rol: "" });
       } catch (error) {
         console.error("Error al crear rol:", error);
       }
@@ -105,7 +109,11 @@ export const RolForm = () => {
   };
 
   const handleMultipleDelete = async (ids) => {
-    if (window.confirm("¿Estás seguro de que quieres eliminar los roles seleccionados?")) {
+    if (
+      window.confirm(
+        "¿Estás seguro de que quieres eliminar los roles seleccionados?"
+      )
+    ) {
       try {
         const response = await fetch(`${apiUrl}${port}/${form}/m`, {
           method: "POST",
@@ -114,39 +122,46 @@ export const RolForm = () => {
           },
           body: JSON.stringify(ids),
         });
-        setDeleted2(true)
-      const data = await response.json();
-
-      data.error == "Ocurrió un error en el servidor" ?  alert("no se puede eliminar depende de un registro") : alert(data.message)
-        
-      } catch (error) {
-        alert(error)
-      }
-  }
-  }
-
-  const handleDelete = async (id) => {
-    if (window.confirm("¿Estás seguro de que quieres eliminar este rol?")) {
-      try {
-        // Enviar solicitud para eliminar un rol existente
-        const response = await fetch(`${apiUrl}${port}/${form}/${id}`, {
-          method: "DELETE",
-        });
+        setDeleted2(true);
         const data = await response.json();
-        data.error == "Ocurrió un error en el servidor" ?  alert("no se puede eliminar depende de un registro") : alert(data.message)
-        
-        // Actualizar el estado después de la eliminación
-        setForm((prev) => prev.filter((info) => info.id != id));
-        setDeleted(true)
-        if (selected && selected.id == id) {
-          setSelected(null);
-          setNew({ id_rol: "", description_rol: "" });
-        }
+
+        data.error == "Ocurrió un error en el servidor"
+          ? alert("no se puede eliminar depende de un registro")
+          : alert(data.message);
       } catch (error) {
-        console.error("Error al eliminar rol:", error);
+        alert(error);
       }
     }
-  };  
+  };
+
+  const handleDelete = async (id) => {
+    if (id > 2) {
+      if (window.confirm("¿Estás seguro de que quieres eliminar este rol?")) {
+        try {
+          // Enviar solicitud para eliminar un rol existente
+          const response = await fetch(`${apiUrl}${port}/${form}/${id}`, {
+            method: "DELETE",
+          });
+          const data = await response.json();
+          data.error == "Ocurrió un error en el servidor"
+            ? alert("no se puede eliminar depende de un registro")
+            : alert(data.message);
+
+          // Actualizar el estado después de la eliminación
+          setForm((prev) => prev.filter((info) => info.id != id));
+          setDeleted(true);
+          if (selected && selected.id == id) {
+            setSelected(null);
+            setNew({ id_rol: "", description_rol: "" });
+          }
+        } catch (error) {
+          console.error("Error al eliminar rol:", error);
+        }
+      }
+    } else {
+      alert("No esta permitido para los roles por defecto.");
+    }
+  };
 
   const handleEdit = (info) => {
     // Cargar los datos del rol seleccionado en los campos de edición
@@ -170,62 +185,78 @@ export const RolForm = () => {
 
   return (
     <>
-    <NavAdmin></NavAdmin>
-    <div className="container mt-4">
-      <div className="row">
-        <div className="col">
-          <TitleComponent title={title.current} size="h1" />
+      <NavAdmin></NavAdmin>
+      <div className="container mt-4">
+        <div className="row">
+          <div className="col">
+            <TitleComponent title={title.current} size="h1" />
+          </div>
         </div>
-      </div>
-      <div className="row">
-        <div className="col-12 col-sm-12 col-md-12 col-lg-6 col-xl-6">
-          <form onSubmit={handleSubmit} className="mb-4">
-            <div className="form-row">
-              <MyInputComponent
-                type="text"
-                required
-                label="Descripción de rol *"
-                name="description_rol"
-                value={news.description_rol}
-                onChange={(e) => handleInputChange(e, setNew)}
-              />
-              <div className="form-group col-md-2 d-flex align-items-end">
-                
-                  {selected ? <button type="submit" className="btn btn-primary btn-block"><UpdateIcon /></button> : forms.length <= 2 ? <button type="submit" className="btn btn-primary btn-block"><AddIcon /></button> : <button type="submit" disabled className="btn btn-primary btn-block"><AddIcon /></button>}
-                
-                &nbsp;
-                {selected && (
-                  <button
-                    type="button"
-                    className="btn btn-danger btn-block mt-2"
-                    onClick={handleCancel}
-                  >
-                    <CloseIcon />
-                  </button>
-                )}
+        <div className="row">
+          <div className="col-12 col-sm-12 col-md-12 col-lg-6 col-xl-6">
+            <form onSubmit={handleSubmit} className="mb-4">
+              <div className="form-row">
+                <MyInputComponent
+                  type="text"
+                  required
+                  label="Descripción de rol *"
+                  name="description_rol"
+                  value={news.description_rol}
+                  onChange={(e) => handleInputChange(e, setNew)}
+                />
+                <div className="form-group col-md-2 d-flex align-items-end">
+                  {selected ? (
+                    <button type="submit" className="btn btn-primary btn-block">
+                      <UpdateIcon />
+                    </button>
+                  ) : forms.length <= 2 ? (
+                    <button type="submit" className="btn btn-primary btn-block">
+                      <AddIcon />
+                    </button>
+                  ) : (
+                    <button
+                      type="submit"
+                      disabled
+                      className="btn btn-primary btn-block"
+                    >
+                      <AddIcon />
+                    </button>
+                  )}
+                  &nbsp;
+                  {selected && (
+                    <button
+                      type="button"
+                      className="btn btn-danger btn-block mt-2"
+                      onClick={handleCancel}
+                    >
+                      <CloseIcon />
+                    </button>
+                  )}
+                </div>
               </div>
-            </div>
-          </form>
-        </div>
-        <div className="col-12 col-sm-12 col-md-12 col-lg-6 col-xl-6">
-          <MyTableComponent
-            data={current}
-            handleDelete={handleDelete}
-            handleEdit={handleEdit}
-            handleMultipleDelete={handleMultipleDelete}
-            idField={"id_rol"}
-            Fields={["description_rol", "createdAt", "updatedAt"]}
-          />
-          <MyPaginationComponent
-            data={forms}
-            PerPage={PerPage}
-            handlePageChange={handlePageChange}
+            </form>
+          </div>
+          <div className="col-12 col-sm-12 col-md-12 col-lg-6 col-xl-6">
+            <MyTableComponent
+              data={current}
+              handleDelete={handleDelete}
+              handleEdit={handleEdit}
+              hiddenDelete={false}
+              hiddenEdit={false}
+              handleMultipleDelete={handleMultipleDelete}
+              idField={"id_rol"}
+              Fields={["description_rol", "createdAt", "updatedAt"]}
             />
-          <br />
-          <br />
+            <MyPaginationComponent
+              data={forms}
+              PerPage={PerPage}
+              handlePageChange={handlePageChange}
+            />
+            <br />
+            <br />
+          </div>
         </div>
       </div>
-    </div>
     </>
   );
 };
