+++ React/project_arduino/src/components/Form/StatusForm.jsx
@@ -7,9 +7,8 @@ import UpdateIcon from "@mui/icons-material/Update";
 import CloseIcon from "@mui/icons-material/Close";
 import AddIcon from "@mui/icons-material/Add";
 import { textRegex, trimRegex } from "../../utils/regex";
-import { apiUrl, port } from '../../utils/config';
+import { apiUrl, port } from "../../utils/config";
 import { NavAdmin } from "../Layouts/NavBar/NavBarAdmin/NavBar";
-
 
 export const StatusForm = () => {
   const [forms, setForm] = useState([]);
@@ -20,11 +19,9 @@ export const StatusForm = () => {
 
   const [currentPage, setCurrentPage] = useState(0); // Estado para el número de página actual
   const PerPage = 6; // Número de roles por página
-  const form = 'status';
+  const form = "status";
 
-  
   const handleGet = async () => {
-    
     try {
       const response = await fetch(`${apiUrl}${port}/${form}`);
       const data = await response.json();
@@ -33,7 +30,7 @@ export const StatusForm = () => {
       console.error("Error al obtener estados:", error);
     }
   };
-  
+
   useEffect(() => {
     // Actualizar la tabla cuando se actualiza el estado 'forms'
     handleGet();
@@ -43,16 +40,15 @@ export const StatusForm = () => {
   useEffect(() => {
     // Actualizar la tabla cuando se actualiza el estado 'forms'
     handleGet();
-    setDeleted(false)
+    setDeleted(false);
   }, [deleted]);
-  
+
   useEffect(() => {
     // Actualizar la tabla cuando se actualiza el estado 'forms'
     handleGet();
-    setDeleted2(false)
+    setDeleted2(false);
   }, [deleted2]);
 
-  
   const handleInputChange = (e) => {
     const { name, value } = e.target;
     if (value.match(textRegex) && !value.match(trimRegex)) {
@@ -62,25 +58,33 @@ export const StatusForm = () => {
 
   const handleSubmit = async (e) => {
     e.preventDefault();
-    
+
     if (!news.description_status) {
       //funciona en el caso del que el campo no sea required
       alert("Por favor, ingrese una descripción válida.");
       return;
     }
-  
+
     if (selected) {
-      if (window.confirm("¿Estás seguro de que quieres actualizar este estado?")) {
+      if (
+        window.confirm("¿Estás seguro de que quieres actualizar este estado?")
+      ) {
         try {
           // Enviar solicitud para actualizar un rol existente
-          const response = await fetch(`${apiUrl}${port}/${form}/${selected.id_status}`, {
-            method: "PUT",
-            headers: {
-              "Content-Type": "application/json",
-            },
-            body: JSON.stringify(news),
-          });
+          const response = await fetch(
+            `${apiUrl}${port}/${form}/${selected.id_status}`,
+            {
+              method: "PUT",
+              headers: {
+                "Content-Type": "application/json",
+              },
+              body: JSON.stringify(news),
+            }
+          );
           const data = await response.json();
+          data.error == "Ocurrió un error en el servidor"
+            ? alert("Ya existe otro estado llamado igual.")
+            : alert(data.message);
           setForm((prev) =>
             prev.map((rol) => (rol.id_status == data.id_status ? data : rol))
           );
@@ -101,8 +105,11 @@ export const StatusForm = () => {
           body: JSON.stringify(news),
         });
         const data = await response.json();
-        
-        data.error == "Ocurrió un error en el servidor" ?  alert("ya existe") : setForm((prev) => [...prev, data]); setNew({ description_status: "" });
+
+        data.error == "Ocurrió un error en el servidor"
+          ? alert("Ya existe otro estado llamado igual.")
+          : setForm((prev) => [...prev, data]);
+        setNew({ description_status: "" });
       } catch (error) {
         console.error("Error al crear rol:", error);
       }
@@ -110,7 +117,11 @@ export const StatusForm = () => {
   };
 
   const handleMultipleDelete = async (ids) => {
-    if (window.confirm("¿Estás seguro de que quieres eliminar los estados seleccionados?")) {
+    if (
+      window.confirm(
+        "¿Estás seguro de que quieres eliminar los estados seleccionados?"
+      )
+    ) {
       const response = await fetch(`${apiUrl}${port}/${form}/m`, {
         method: "POST",
         headers: {
@@ -118,39 +129,52 @@ export const StatusForm = () => {
         },
         body: JSON.stringify(ids),
       });
-      setDeleted2(true)
-    const data = await response.json();
-    data.error == "Ocurrió un error en el servidor" ?  alert("no se puede eliminar depende de un registro") : alert(data.message)
-  }
-  }
+      setDeleted2(true);
+      const data = await response.json();
+      data.error == "Ocurrió un error en el servidor"
+        ? alert("no se puede eliminar depende de un registro")
+        : alert(data.message);
+    }
+  };
 
   const handleDelete = async (id) => {
-    if (window.confirm("¿Estás seguro de que quieres eliminar este estado?")) {
-      try {
-        // Enviar solicitud para eliminar un rol existente
-        const response = await fetch(`${apiUrl}${port}/${form}/${id}`, {
-          method: "DELETE",
-        });
-        const data = await response.json();
-        data.error == "Ocurrió un error en el servidor" ?  alert("no se puede eliminar depende de un registro") : alert(data.message)
-        
-        // Actualizar el estado después de la eliminación
-        setForm((prev) => prev.filter((info) => info.id != id));
-        setDeleted(true)
-        if (selected && selected.id == id) {
-          setSelected(null);
-          setNew({ id_status: "", description_status: "" });
+    if (id > 2) {
+      if (
+        window.confirm("¿Estás seguro de que quieres eliminar este estado?")
+      ) {
+        try {
+          // Enviar solicitud para eliminar un rol existente
+          const response = await fetch(`${apiUrl}${port}/${form}/${id}`, {
+            method: "DELETE",
+          });
+          const data = await response.json();
+          data.error == "Ocurrió un error en el servidor"
+            ? alert("no se puede eliminar depende de un registro")
+            : alert(data.message);
+
+          // Actualizar el estado después de la eliminación
+          setForm((prev) => prev.filter((info) => info.id != id));
+          setDeleted(true);
+          if (selected && selected.id == id) {
+            setSelected(null);
+            setNew({ id_status: "", description_status: "" });
+          }
+        } catch (error) {
+          console.error("Error al eliminar rol:", error);
         }
-      } catch (error) {
-        console.error("Error al eliminar rol:", error);
       }
+    } else {
+      alert("No esta permitido para los estados por defecto.");
     }
-  };  
+  };
 
   const handleEdit = (info) => {
     // Cargar los datos del rol seleccionado en los campos de edición
     setSelected(info);
-    setNew({ id_status: info.id_status, description_status: info.description_status });
+    setNew({
+      id_status: info.id_status,
+      description_status: info.description_status,
+    });
   };
 
   const handleCancel = () => {
@@ -169,61 +193,76 @@ export const StatusForm = () => {
 
   return (
     <>
-    <NavAdmin></NavAdmin>
-    <div className="container mt-4">
-      <div className="row">
-        <div className="col">
-          <TitleComponent title="Administrador de estados" size="h1" />
+      <NavAdmin></NavAdmin>
+      <div className="container mt-4">
+        <div className="row">
+          <div className="col">
+            <TitleComponent title="Administrador de estados" size="h1" />
+          </div>
         </div>
-      </div>
-      <div className="row">
-        <div className="col-12 col-sm-12 col-md-12 col-lg-5 col-xl-5">
-          <form onSubmit={handleSubmit} className="mb-4">
-            <div className="form-row">
-              <MyInputComponent
-                type="text"
-                required
-                label="Descripción del estado *"
-                name="description_status"
-                value={news.description_status}
-                onChange={handleInputChange}
-              />
-              <div className="form-group col-md-2 d-flex align-items-end">
-              {selected ? <button type="submit" className="btn btn-primary btn-block"><UpdateIcon /></button> : forms.length < 2 ? <button type="submit" className="btn btn-primary btn-block"><AddIcon /></button> : <button type="submit" disabled className="btn btn-primary btn-block"><AddIcon /></button>}
-
-                &nbsp;
-                {selected && (
-                  <button
-                    type="button"
-                    className="btn btn-danger btn-block mt-2"
-                    onClick={handleCancel}
-                  >
-                    <CloseIcon />
-                  </button>
-                )}
+        <div className="row">
+          <div className="col-12 col-sm-12 col-md-12 col-lg-5 col-xl-5">
+            <form onSubmit={handleSubmit} className="mb-4">
+              <div className="form-row">
+                <MyInputComponent
+                  type="text"
+                  required
+                  label="Descripción del estado *"
+                  name="description_status"
+                  value={news.description_status}
+                  onChange={handleInputChange}
+                />
+                <div className="form-group col-md-2 d-flex align-items-end">
+                  {selected ? (
+                    <button type="submit" className="btn btn-primary btn-block">
+                      <UpdateIcon />
+                    </button>
+                  ) : forms.length < 2 ? (
+                    <button type="submit" className="btn btn-primary btn-block">
+                      <AddIcon />
+                    </button>
+                  ) : (
+                    <button
+                      type="submit"
+                      disabled
+                      className="btn btn-primary btn-block"
+                    >
+                      <AddIcon />
+                    </button>
+                  )}
+                  &nbsp;
+                  {selected && (
+                    <button
+                      type="button"
+                      className="btn btn-danger btn-block mt-2"
+                      onClick={handleCancel}
+                    >
+                      <CloseIcon />
+                    </button>
+                  )}
+                </div>
               </div>
-            </div>
-          </form>
-        </div>
-        <div className="col-12 col-sm-12 col-md-12 col-lg-7 col-xl-7">
-          <MyTableComponent
-            data={current}
-            handleDelete={handleDelete}
-            handleEdit={handleEdit}
-            handleMultipleDelete={handleMultipleDelete}
-            idField={"id_status"}
-            Fields={["description_status", "createdAt", "updatedAt"]}
-          />
-          <MyPaginationComponent
-            data={forms}
-            PerPage={PerPage}
-            handlePageChange={handlePageChange}
-          />
-          <br />
-          <br />
+            </form>
+          </div>
+          <div className="col-12 col-sm-12 col-md-12 col-lg-7 col-xl-7">
+            <MyTableComponent
+              data={current}
+              handleDelete={handleDelete}
+              handleEdit={handleEdit}
+              handleMultipleDelete={handleMultipleDelete}
+              idField={"id_status"}
+              Fields={["description_status", "createdAt", "updatedAt"]}
+            />
+            <MyPaginationComponent
+              data={forms}
+              PerPage={PerPage}
+              handlePageChange={handlePageChange}
+            />
+            <br />
+            <br />
+          </div>
         </div>
       </div>
-    </div>
     </>
   );
 };
